backend/requirements.txt
fastapi
uvicorn[standard]
google-play-scraper
httpx
python-dotenv
pydantic
